Drop new game code
That was a second attempt to refactor the code to make it more parallel
and testable.

Unfortunately, I don't think I have it in me to finish this rewrite, so
it's best to revert to the stable code.

Also DDD architecture I tried to apply here doesn't *really* fit a game
engine. Event bus was rather out of place here. It is a good idea
however to isolate the game logic from the infrastructure code.